backend/pyproject.toml
[project]
name = "interview-pilot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "amazon-transcribe>=0.6.4",
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.54",
    "ddgs>=9.10.0",
    "fastapi>=0.129.2",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "fpdf2>=2.8.6",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
